* commit 'b8497eb6c046f836541f69167239bdb97498ace2': (57 commits)
  Removing simulation process management
  Adding a "from script" option for MC position control sim
  Adding stand-to-crawl script with box
  redoing leaning back script
  Add standing and trying to balance on one toe and two arms.
  Add trying to get down script.
  Multi-contact script reverser
  Adding script leaning back on tilted handholds
  Adding multi-contact environments with handholds
  Changing method signature for PlanarRegionTools distance calculation to take read-only point
  Removing ContactStateChangeMessage
  Extracting multi-contact environments from position control simulation
  Add another initial robot pose.
  Standing to getting down on the side boxes.
  Adding script using box in front
  Boxes for help getting down.
  Adding new multi-contact scripts using a box
  Adding some new multi-contact environments
  Adding new pelvis near level script
  A script that goes from standing to down on all fours and takes a few hand steps but topples over.
  ...

+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/gdx/ui/missionControl/RestartableMissionControlProcess.java
@@ -14,11 +14,13 @@ public abstract class RestartableMissionControlProcess implements MissionControl
    volatile boolean started = false;
    volatile boolean stopping = false;
 
-   private final ResettableExceptionHandlingExecutorService executorService;
+   private ResettableExceptionHandlingExecutorService executorService;
 
-   public RestartableMissionControlProcess()
+   public ResettableExceptionHandlingExecutorService getOrCreateExecutor()
    {
-      executorService = MissingThreadTools.newSingleThreadExecutor(getName() + "ManagementThread", true);
+      if (executorService == null)
+         executorService = MissingThreadTools.newSingleThreadExecutor(getName() + "ManagementThread", true);
+      return executorService;
    }
 
    protected abstract void startInternal();
@@ -65,7 +67,7 @@ public abstract class RestartableMissionControlProcess implements MissionControl
    public void start()
    {
       starting = true;
-      executorService.execute(() ->
+      getOrCreateExecutor().execute(() ->
       {
          startInternal();
          starting = false;
@@ -77,7 +79,7 @@ public abstract class RestartableMissionControlProcess implements MissionControl
    {
       started = false;
       stopping = true;
-      executorService.execute(() ->
+      getOrCreateExecutor().execute(() ->
       {
          stopInternal();
          stopping = false;
@@ -121,7 +123,7 @@ public abstract class RestartableMissionControlProcess implements MissionControl
             LogTools.error("{} ran out of time to stop!", getName());
          }
       }
-      if (!executorService.isShutdown())
+      if (executorService != null && !executorService.isShutdown())
          executorService.destroy();
    }
 }
